Uploaded car.sql, the SQLite version
I demonstrate object oriented programming in SQLite using my car blueprint that I use in some of my other projects. SQLite is a version of SQL that, as its name suggests, is a "lite" version. Therefore, unlike MySQL, PostgreSQL, SQL Server, MariaDB, and Access, it does not support stored procedures as of SQLite3. Therefore, "functions" cannot be created, and standard SQL in the SQLite flavor must be written in order to commit changes.